# Do not send field query when using reserved connection