--- a/go/vt/vttablet/tabletserver/fuzz.go
+++ b/go/vt/vttablet/tabletserver/fuzz.go
@@ -67,6 +67,6 @@ func FuzzGetPlan(data []byte) int {
 	qe.SetQueryPlanCacheCap(1024)
 
 	// Call target
-	_, _ = qe.GetPlan(context.Background(), logStats, query2, true, false)
+	_, _ = qe.GetPlan(context.Background(), logStats, query2, true, 0, nil)
 	return 1
 }