--- a/go/vt/vttablet/tabletserver/query_engine.go
+++ b/go/vt/vttablet/tabletserver/query_engine.go
@@ -25,6 +25,8 @@ import (
 	"sync/atomic"
 	"time"
 
+	"vitess.io/vitess/go/sqltypes"
+
 	"context"
 
 	"vitess.io/vitess/go/acl"
@@ -287,7 +289,7 @@ func (qe *QueryEngine) Close() {
 }
 
 // GetPlan returns the TabletPlan that for the query. Plans are cached in a cache.LRUCache.
-func (qe *QueryEngine) GetPlan(ctx context.Context, logStats *tabletenv.LogStats, sql string, skipQueryPlanCache bool, isReservedConn bool) (*TabletPlan, error) {
+func (qe *QueryEngine) GetPlan(ctx context.Context, logStats *tabletenv.LogStats, sql string, skipQueryPlanCache bool, reservedConnID int64, te *TxEngine) (*TabletPlan, error) {
 	span, ctx := trace.NewSpan(ctx, "QueryEngine.GetPlan")
 	defer span.Finish()
 
@@ -308,24 +310,39 @@ func (qe *QueryEngine) GetPlan(ctx context.Context, logStats *tabletenv.LogStats
 	if err != nil {
 		return nil, err
 	}
-	splan, err := planbuilder.Build(statement, qe.tables, isReservedConn, qe.env.Config().DB.DBName)
+	splan, err := planbuilder.Build(statement, qe.tables, reservedConnID != 0, qe.env.Config().DB.DBName)
 	if err != nil {
 		return nil, err
 	}
 	plan := &TabletPlan{Plan: splan, Original: sql}
 	plan.Rules = qe.queryRuleSources.FilterByPlan(sql, plan.PlanID, plan.TableNames()...)
 	plan.buildAuthorized()
-	if !isReservedConn && plan.PlanID.IsSelect() {
+	if plan.PlanID.IsSelect() {
 		if !skipQueryPlanCache && qe.enableQueryPlanFieldCaching && plan.FieldQuery != nil {
-			conn, err := qe.conns.Get(ctx)
-			if err != nil {
-				return nil, err
+			var statefulConn *StatefulConnection
+			var conn *connpool.DBConn
+			if reservedConnID != 0 {
+				statefulConn, err = te.txPool.GetAndLock(reservedConnID, "")
+				if err != nil {
+					return nil, err
+				}
+				defer statefulConn.Unlock()
+			} else {
+				conn, err = qe.conns.Get(ctx)
+				if err != nil {
+					return nil, err
+				}
+				defer conn.Recycle()
 			}
-			defer conn.Recycle()
 
 			sql := plan.FieldQuery.Query
 			start := time.Now()
-			r, err := conn.Exec(ctx, sql, 1, true)
+			var r *sqltypes.Result
+			if reservedConnID != 0 {
+				r, err = statefulConn.Exec(ctx, sql, 1, true)
+			} else {
+				r, err = conn.Exec(ctx, sql, 1, true)
+			}
 			logStats.AddRewrittenSQL(sql, start)
 			if err != nil {
 				return nil, err