## Description

This Pull Request changes the query engine so it does not send the field query for `SELECT` statement when we are in a reserved connection. This is done so we can use the same reserved connection as the actual query when sending the field query, and thus, use the same system variables, like `sql_mode` for instance.

## Checklist

-   [x] "Backport me!" label has been added if this change should be backported
-   [x] Tests were added or are not required
-   [x] Documentation was added or is not required