@harshit-gangal 2022-04-28
I think we should benchmark this change to see the impact of not catching fields.