@frouioui 2022-04-28
@harshit-gangal I added a benchmark in 2434d489eb.

The results without the change are:

```
goos: darwin
goarch: amd64
pkg: vitess.io/vitess/go/test/endtoend/vtgate/reservedconn
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkReservedConnFieldQuery-16          1230           1284618 ns/op
BenchmarkReservedConnFieldQuery-16           790           1292526 ns/op
BenchmarkReservedConnFieldQuery-16           984           1366862 ns/op
BenchmarkReservedConnFieldQuery-16           807           1613920 ns/op
BenchmarkReservedConnFieldQuery-16           412           2968577 ns/op
BenchmarkReservedConnFieldQuery-16           586           2412719 ns/op
BenchmarkReservedConnFieldQuery-16           478           2917387 ns/op
BenchmarkReservedConnFieldQuery-16           494           2634667 ns/op
BenchmarkReservedConnFieldQuery-16           463           2602573 ns/op
BenchmarkReservedConnFieldQuery-16           572           2096607 ns/op
PASS
```

And here are the results with the change:

```
goos: darwin
goarch: amd64
pkg: vitess.io/vitess/go/test/endtoend/vtgate/reservedconn
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkReservedConnFieldQuery-16           540           2279100 ns/op
BenchmarkReservedConnFieldQuery-16           415           2712610 ns/op
BenchmarkReservedConnFieldQuery-16           459           2423340 ns/op
BenchmarkReservedConnFieldQuery-16           454           2336300 ns/op
BenchmarkReservedConnFieldQuery-16           530           2026756 ns/op
BenchmarkReservedConnFieldQuery-16           572           1803834 ns/op
BenchmarkReservedConnFieldQuery-16           584           1748262 ns/op
BenchmarkReservedConnFieldQuery-16           613           1816961 ns/op
BenchmarkReservedConnFieldQuery-16           589           1713665 ns/op
BenchmarkReservedConnFieldQuery-16           614           1758979 ns/op
PASS
```

And here is the `benchcmp`:

```
benchmark                              old ns/op     new ns/op     delta
BenchmarkReservedConnFieldQuery-16     1284618       2279100       +77.41%
BenchmarkReservedConnFieldQuery-16     1292526       2712610       +109.87%
BenchmarkReservedConnFieldQuery-16     1366862       2423340       +77.29%
BenchmarkReservedConnFieldQuery-16     1613920       2336300       +44.76%
BenchmarkReservedConnFieldQuery-16     2968577       2026756       -31.73%
BenchmarkReservedConnFieldQuery-16     2412719       1803834       -25.24%
BenchmarkReservedConnFieldQuery-16     2917387       1748262       -40.07%
BenchmarkReservedConnFieldQuery-16     2634667       1816961       -31.04%
BenchmarkReservedConnFieldQuery-16     2602573       1713665       -34.15%
BenchmarkReservedConnFieldQuery-16     2096607       1758979       -16.10%
```

The average delta is `+13.1%`, so from these observations, the new code path is about `13.1%` slower.