@harshit-gangal 2022-04-29
I think we should block those sql_modes or system variables that can lead to different fields (with and without reserved connection) so that we can cache it.