@frouioui 2022-05-02
@harshit-gangal, the micro-benchmark comparison between A and B, with A being this PR without the reusable reserved connections (f38bc2a) and the unsupported sql_mode check (a67b92d), and B being this PR's HEAD, are as follows:

```
benchmark                              old ns/op     new ns/op     delta
BenchmarkReservedConnFieldQuery-16     2127361       1944415       -8.60%
BenchmarkReservedConnFieldQuery-16     1416054       1292569       -8.72%
BenchmarkReservedConnFieldQuery-16     1341221       1306599       -2.58%
BenchmarkReservedConnFieldQuery-16     1449124       1272365       -12.20%
BenchmarkReservedConnFieldQuery-16     1479808       1302996       -11.95%
BenchmarkReservedConnFieldQuery-16     1370051       1394855       +1.81%
BenchmarkReservedConnFieldQuery-16     1408754       1553413       +10.27%
BenchmarkReservedConnFieldQuery-16     1914588       1691912       -11.63%
BenchmarkReservedConnFieldQuery-16     1271019       1484090       +16.76%
BenchmarkReservedConnFieldQuery-16     1263342       1493031       +18.18%
```

Overall there is not a lot of difference, the results are pretty similar.