@frouioui 2022-05-05
Comparison between: **before the feature** _and_ **after the feature**:
```
name                       old time/op  new time/op  delta
ReservedConnFieldQuery-48  2.13ms ± 4%  2.16ms ± 3%  +1.26%  (p=0.039 n=19+18)
```

Comparison between: **with the feature** _and_ **with feature + unauthorized sql_mode and reusable reserved connections**:
```
name                       old time/op  new time/op  delta
ReservedConnFieldQuery-48  2.16ms ± 3%  2.11ms ± 6%  -2.53%  (p=0.006 n=18+20)
```

The p-value that we got for both comparisons seems low enough to me to confirm that this change is not negatively impacting the performance of Vitess.